pyproject.toml
[tool.poetry]
name = "fitsync_backend"
version = "0.1.0"
description = ""
authors = ["Daniel Montilla <danielmontillanavas@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.4"}
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
mypy = "^1.13.0"
pre-commit = "^4.0.1"
types-requests = "^2.32.0.20241016"
pytest-asyncio = "^0.24.0"

[tool.mypy]
ignore_missing_imports = true
show_column_numbers = true
disallow_incomplete_defs = true
namespace_packages = true

[tool.ruff]
target-version = "py311"
line-length = 120
exclude = [
    ".git", ".ipynb_checkpoints", ".mypy_cache", ".pyenv", ".pytest_cache", ".ruff_cache",
    ".venv", "__pypackages__", ".idea", "__pycache__", "htmlcov", "dist", '*/prompts/*'
]
respect-gitignore = true

[tool.ruff.lint]
fixable = ["ALL"]
select = ["A", "E", "F", "W", "N", "B", "I001", "ERA001"]
ignore = ["B008"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-wrap-aliases = true
lines-after-imports = -1 # automatic setup
relative-imports-order = "closest-to-furthest"
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # Module imported but not used

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
